Revert PR changes (check 20.20.3 again)
